+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 1.1.0
 
+- Prismarine support
 - Added `delete` command
+- Added `cleanup` command
+- Added `inspect` command
 
 ## 1.0.0